last minute commit

+++ level.cpp
@@ -4,7 +4,7 @@ Level::Level()
 {
 }
 
-Level::Level(int level_number, std::vector<int> zombie_sequence, int row_number, int start_time, int interval, int interval_decrement)
+Level::Level(int level_number, std::vector<int> zombie_sequence, int row_number, int start_time, int interval, double interval_decrement)
 {
     levelNumber = level_number;
     zombieSequence = zombie_sequence;
